add error handling and ui states

+++ data/src/main/java/com/varosyan/data/db/dao/UserDao.kt
@@ -19,7 +19,7 @@ interface UserDao {
   """
     )
     suspend fun getUsersFromId(
-        startId: Int,
+        startId: Long,
         count: Int
     ): List<UserEntity>
